refactor: move http helper to a folder
