+++ src/presentation/controllers/login/index.ts
@@ -7,7 +7,7 @@ import {
   internalException,
   ok,
   unauthorized,
-} from '@/presentation/helpers/http'
+} from '@/presentation/helpers/http/http'
 
 import {
   IController,
